Add a bit of docs
