+++ Letterbook.Api/Program.cs
@@ -13,7 +13,7 @@ public class Program
 
         builder.Services.AddControllers(options =>
         {
-            options.Conventions.Add(new RouteTokenTransformerConvention(new RouteTransformer()));
+            options.Conventions.Add(new RouteTokenTransformerConvention(new SnakeCaseRouteTransformer()));
         });
         // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
         builder.Services.AddEndpointsApiExplorer();
